Subdivision: basic GPU subdivision for OpenGL
This implements GPU subdivision via OpenSubDiv for OpenGL based render
engines. At the moment only uniform subdivision is supported. Adaptive
subdivision still lacks a way to render triangles from the subdivision
patches.

The GPUBatch is extended with a new VBO and a new IBO to gather
subdivided points and indices. This allows to render the subdivided
surface and the original control cage in edit mode without disrupting
too much the code.

To generate triangles from the subdivided index buffer, a compute
shader is used. Although, at the moment normals are not computed there
so a CPU fallback is used, where data is read from the GPU, processed
on the CPU to generate triangles and normals, and then sent back to
the GPU. To compute normals, a secondary compute shader might be used.
This is not ideal for performance, but allows to validate that
subdivision is indeed performed properly on the GPU.

Attributes (including UVs and vertex colors) are not supported yet.

For performance, some proper caching mechanism will need to be used.
There currently is some logic to cache the subdivision structures that
can be reused between updates. However, it is deactivated as it does
not properly detect when changes occur, and causes crashes in edit
mode.

Performance can be further improved by reducing data transfer to the
GPU, since OpenSubDiv asks us to allocate buffers big enough for the
refined and coarse mesh we also allocate such buffer on the host, and
transfer the entier buffer worth of data to the device, although only
the coarse part need to be truly allocated on the host and sent to the
device. This will require some change to the vertex buffer structure.