VSE UX: Make Speed Effect strips more user friendly.
The `Speed Effect` panel is confusing. There are many toggles and hidden values to control the speed effect.
So this patch proposes to simplify it, by adding an `enum` and making it more intuitive by drawing f-curves on the strips.

Before:
{F7843850}

After:
{F9775464,size=full}

**Enum Changes:**
- The new enums correspond to 4 modes: `Stretch`, `Multiply`, `Frame Number` and `Length`.
- Except `Stretch` the other modes now has its respective control values.

**Drawing Changes:**
- F-curve drawing for Stretch, Multiply, Length and Frame Number.
- Value drawing when no keyframes for Stretch, Length and Frame Numbers.

General view of the new drawing for each speed effect mode:
{F9796642, size=full}

Detail of the horizontal zero (blue) line in the new `Multiply` mode:
{F9798520, size=full}

Nice to have (but I don't know how):
- Auto adjusting of endframe when using Multiply or Boost.

---
**OLD STUFF FOR REFERENCE:**
The current functions are that confusing that I feel it is necessary to "spell" each one of them out:

Use Default Fade On - Was renamed to "Stretch to input strip length" in 2.8+ because the adjusts the speed so the first and last frame is the same no matter what the duration of the strip is. Default Fade makes no sense, imo. In the enum I would call it "Fit Duration" or "Stretch to Duration", but maybe it'll be too long? As the speed is relative to the duration of the clip the Multiplier has no meaning and should be disabled, imo. The tooltip could be changed to something along the lines of: "Stretch to input strip duration" or "Fit speed to input strip duration"(gif):
{F8488797}

The Use as Speed does exactly the same as the Multiply Speed setting, making the latter redundant, imo. It could be called "Speed" in the enum, and the value "Factor". In the tooltip "or remapping the frame number" should be removed. How the Use Speed is working(gif):
{F8488800}

When "Scale to Length" is selected the strip duration goes from 0%(first frame) to 100%(last frame), therefore doesn't it make sense to call the value "Frame Number" but should be called "Percentage" and the Value should be set to go from 0 to 100. "Scale to Length" is not very informative, and since it only works when animated. "Multiply" has no function here either.
How it is working(gif):
{F8488801}

The fourth setting, yes, there are four, is when "Scale to Length" is unselected, and behaves like the previous setting, but in frame numbers.

Differential Revision: https://developer.blender.org/D6110