Support opening and collapsing tree-view items
For this to work, we need to support tree-view items with state that is
persistent over redraws. This is achieved by matching tree items to tree
items of a previous redraw, and copying the state if needed. This is
done in the tree-view API, so except for more advanced cases (with
custom item state or custom ways to match items), API users shouldn't
have to deal with this.