Asset Catalogs: move from string identifier to UUIDs
Move from a string `catalog_id` to a UUID. This avoids having to keep
the ID of the catalog in sync with its semantics (so renaming from X to
Y can keep the UUID the same).

Since UUIDs don't communicate any human-readable information, the
mapping from catalog UUID to its path (stored in the Catalog Definition
File, CDF) is critical for understanding which asset is stored in which
human-readable catalog. To make this less critical, and to allow manual
data reconstruction after a CDF is lost/corrupted, each catalog also has
a "simple name" that's stored along with the UUID. This is also stored
on each asset, next to the catalog UUID.