Asset Catalogs: implement catalog deletion
Asset catalogs can now be deleted by providing their UUID. This marks the
catalog for deletion, which causes it to be hidden from the UI and not
saved to disk.

The catalog is not immediately deleted from memory, to make it possible
to later merge on-disk data with in-memory data. This is envisioned to
support multi-user edits to the same catalog definition file (for example
when synced between machines with SyncThing).