Cleanup: refactor asset catalog definition file loading
Change the API for loading asset catalog definition files, by moving the
file/line parsing code into the `AssetCatalogDefinitionFile` class.

This separates the parsing of these files from handling the parsed data.
The caller of the `parse_catalog_line()` function has to provide a callback
that stores the `AssetCatalog` in the desired place.

No functional changes.