Use existing C-APIs for path and filesystem handling
The `ghc::filesystem` we wanted to use for a platform compatible
replacement of `std::filesystem` (see D12197) doesn't behave as we
expect or want to. We could probably find a solution, but it's not worth
the extra time. To move this forward, we can switch the asset catalogs
to use Blender's C functions for path and filesystem handling.

Differential Revision: https://developer.blender.org/D12538