Support deleting catalogs from the UI
Adds a 'x' icon to catalogs to delete them (including children, if any).

There's no undo/redo for editing catalogs yet (since they are not
stored in .blend files), so deleting a catalog is a destructive action.
Therefore we show a confirmation popup.

Only catalog tree-items in the UI that represent an actual catalog can
be deleted for now. With T91451 done, this will change (every item
visible will be an actual catalog, so you'll be able to delete each).